training/jersey-numbers/requirements.txt
# 核心依賴
ultralytics>=8.0.0
roboflow>=1.1.0
opencv-python>=4.8.0
pillow>=9.5.0
numpy>=1.24.0
pyyaml>=6.0

# 訓練相關
# 注意：PyTorch 需要根據 CUDA 版本單獨安裝
# 請參考 README.md 中的安裝步驟
# torch>=2.0.0
# torchvision>=0.15.0

# 資料處理
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 可選：TensorBoard 用於訓練監控
# tensorboard>=2.13.0
